+++ example/src/App.svelte
@@ -2,7 +2,7 @@
 	import Typewriter from 'Typewriter.svelte'
 </script>
 
-<Typewriter loopRandom on:done={() => console.log('Hello World!')}>
+<Typewriter cascade on:done={() => console.log('Hello World!')}>
   <div>some text...</div>
   <div>Foo</div>
   <a href="user">some link</a>