+++ src/modes/cascade.js
@@ -4,12 +4,16 @@ import { typingInterval } from '../utils'
 // Unify the function below with the one on cascade mode
 /** @type {TypewriterEffectFn} */
 const typewriterEffect = async ({ currentNode, text }, options) => {
+	currentNode.classList.add('typing')
 	for (let index = 0; index <= text.length; index++) {
 		const char = text[index]
 		char === '<' && (index = text.indexOf('>', index))
 		currentNode.innerHTML = text.slice(0, index)
 		await typingInterval(options.interval)
 	}
+	currentNode.nextSibling !== null && currentNode.classList.length == 1
+		? currentNode.removeAttribute('class')
+		: currentNode.classList.remove('typing')
 }
 
 /** @type {TypewriterModeFn} */